Correction du calcul de l'airmass et ajout de tests pour vérifier les valeurs d'airmass, ainsi que la conversion de la priorité en entier lors de la planification des observations. Ajout d'un nouveau fichier CLI pour gérer les options de planification des observations.

+++ nextgen/rust_app/src/scheduler.rs
@@ -142,7 +142,7 @@ pub fn schedule_observation_queue(
                 target_name: target.name,
                 ra_deg: target.ra_deg,
                 dec_deg: target.dec_deg,
-                priority: target.priority,
+                priority: target.priority as u32,
                 altitude_deg: vis.altitude_deg,
                 azimuth_deg: vis.azimuth_deg,
                 airmass: vis.airmass,
